fix: type

+++ AreaCalculation/Circle/CircleInfo.cs
@@ -4,5 +4,5 @@ namespace AreaCalculation.Circle;
 
 public class CircleInfo: IFigureInfo
 {
-    public float Radius { get; set; }
+    public double Radius { get; set; }
 }
